move duplicated code utils to shared
