+++ src/renderer/src/components/song/SongImage.tsx
@@ -1,5 +1,5 @@
 import defaultBackground from "@/assets/osu-default-background.jpg";
-import { availableResource, getResourcePath } from "@/lib/tungsten/resource";
+import { availableResource, getResourcePath } from "@shared/lib/tungsten/resource";
 import { Accessor, Component, createEffect, createSignal, JSX, onCleanup, onMount } from "solid-js";
 
 const SET_SOURCE_EVENT = "set-source";
