+++ src/renderer/src/lib/draggable/draggable.ts
@@ -1,5 +1,5 @@
 //todo redo from parent perspective... how did i not think of this... I'm quite stupid
-import { Vec2, vec2Length } from "@/lib/tungsten/math";
+import { Vec2, vec2Length } from "@shared/lib/tungsten/math";
 import defaultHint from "./defaultHint";
 import scrollAnimation from "./scrollAnimation";
