+++ src/shared/lib/rust-types/Result.ts
