+++ src/shared/lib/rusty-types/Optional.ts
