+++ src/shared/lib/rusty-types/Result.ts
