+++ src/shared/lib/tungsten/resource.ts
