+++ tsconfig.web.json
@@ -4,7 +4,8 @@
     "src/renderer/src/env.d.ts",
     "src/renderer/src/**/*",
     "src/renderer/src/**/*.tsx",
-    "src/preload/*.d.ts"
+    "src/preload/*.d.ts",
+    "src/shared/**/*"
   ],
   "compilerOptions": {
     "noImplicitAny": true,
@@ -13,7 +14,8 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["src/renderer/src/*"],
-      "@types": ["src/@types.d.ts"]
+      "@types": ["src/@types.d.ts"],
+      "@shared/*": ["src/shared/*"]
     }
   }
 }
